Fixed MvcFileUpload Bundling Issue
I was unable to get the bundling feature to properly work in the
_MvcFileUpload partial view. One of the problems was the fact that you
cannot bundle multiple scripts from a CDN in the MVC Framework and so
the code that I was using was actually just providing a fallback
location in case the CDN script wasn't available.  The other issue was
that I could not include bundle together the scripts required for file
upload (even though they are all locally hosted...) else i would receive
errors; so I created a separate bundle for each script and loaded them
individually.