Dispatch Job Materials Pulled from Indented BOM
I changed the implementation where a stored procedure was used to look
up the materials for each job-suffix.  Instead I look up the data
directly from the jobmatl table and look only by job so that if a BOM is
indented, all job materials for the job (regardless of job suffix) will
be pulled into the dispatch job material list.  I then implemented the
shell, frame and fabric methods as readonly properties that produced
values based on matches from the jobmatl list for the job.  This should
solve the issue where we have created another level in the BOM.