Implemented Repositories 4 Dispatch & Orders
I removed the extension method that was gathering the remaining orders
and instead moved it into the Repository where all the searching is done
for DataTables.  I have both the HaworthDispatchJob and HaworthOrder
Repositories ready to implement for searching criteria and returning the
appropriate values (record count, filtered records, etc.) to DataTables.