Dialog for holding pdf and IDLabel fields
I was able to get IE to display the ID Label pdf in a dialog box by
using and html object tag. However, when I did this then firefox stopped
displaying the pdf in a dialog.  So I implemented javascript to check
what kind of browser is being used and if it is IE then I use an object
tag, otherwise I use an embed tag.  Ultimately the issue is that when an
embed tag is used to hold a pdf document in IE, the browser aborts the
http GET.
I populated the remaining required fields for the IDLabel and created
another report (rdlc) for the scenario where the appropriate data for an
ID Label isn't found.
I also placed a attribute in web.debug.config and web.release.config so
that when the application gets published, the browser can find the rdlc
files which get placed in the Bin directory.